# openapi/swagger.yaml
basePath: /api/v1
definitions:
  handlers.ApprovePaymentRequest:
    properties:
      paid_amount:
        type: number
    type: object
...
  models.PaymentResponse:
    properties:
      amount:
        type: number
      applicant_identity:
        type: string
      applicant_name:
        type: string
      applicant_phone:
        type: string
      approved_at:
        type: string
      contract_id:
        type: integer
      contract_status:
        description: Contract details
        type: string
      description:
        type: string
      due_date:
        type: string
      has_receipt:
        type: boolean
      id:
        type: integer
      interest_amount:
        type: number
      lot_address:
        type: string
      lot_dimensions:
        type: string
      lot_name:
        type: string
      overdue_days:
        type: integer
      paid_amount:
        type: number
      payment_date:
        type: string
      payment_type:
        type: string
      project_address:
        type: string
      project_name:
        type: string
      status:
        type: string
    type: object
...
  repository.PaymentStats:
    properties:
      collected_this_month:
        type: number
      pending_this_month:
        type: number
      total_overdue:
        type: number
    type: object
...
  services.RevenuePoint:
    properties:
      amount:
        type: number
      date:
        type: string
    type: object
host: localhost:8081
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: REST API for Fintera Real Estate Management System
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Fintera API
  version: "1.0"
paths:
...
  /payments:
    get:
      consumes:
      - application/json
      description: Get a paginated list of payments
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: per_page
        type: integer
      - description: Filter by status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List Payments
      tags:
      - Payments
  /payments/{payment_id}:
    get:
      consumes:
      - application/json
      description: Get a payment by ID
      parameters:
      - description: Payment ID
        in: path
        name: payment_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaymentResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get Payment
      tags:
      - Payments
  /payments/{payment_id}/approve:
    post:
      consumes:
      - application/json
      description: Approve a payment (Admin)
      parameters:
      - description: Payment ID
        in: path
        name: payment_id
        required: true
        type: integer
      - description: Amount
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.ApprovePaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaymentResponse'
      security:
      - BearerAuth: []
      summary: Approve Payment
      tags:
      - Payments
  /payments/{payment_id}/receipt:
    get:
      description: Download payment receipt
      parameters:
      - description: Payment ID
        in: path
        name: payment_id
        required: true
        type: integer
      produces:
      - application/octet-stream
      responses:
        "200":
          description: receipt
          schema:
            type: file
      security:
      - BearerAuth: []
      summary: Download Receipt
      tags:
      - Payments
    post:
      consumes:
      - multipart/form-data
      description: Upload payment receipt image/pdf
      parameters:
      - description: Payment ID
        in: path
        name: payment_id
        required: true
        type: integer
      - description: Receipt File
        in: formData
        name: receipt
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Upload Receipt
      tags:
      - Payments
  /payments/{payment_id}/reject:
    post:
      consumes:
      - application/json
      description: Reject a payment (Admin)
      parameters:
      - description: Payment ID
        in: path
        name: payment_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaymentResponse'
      security:
      - BearerAuth: []
      summary: Reject Payment
      tags:
      - Payments
  /payments/{payment_id}/undo:
    post:
      consumes:
      - application/json
      description: Revert payment state (Admin)
      parameters:
      - description: Payment ID
        in: path
        name: payment_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Undo Payment
      tags:
      - Payments
  /payments/statistics:
    get:
      consumes:
      - application/json
      description: Get payment statistics for revenue flow
      parameters:
      - description: Month (1-12)
        in: query
        name: month
        required: true
        type: integer
      - description: Year (YYYY)
        in: query
        name: year
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/services.RevenuePoint'
            type: array
      security:
      - BearerAuth: []
      summary: Payment Statistics
      tags:
      - Payments
  /payments/stats:
    get:
      consumes:
      - application/json
      description: Get monthly payment statistics (pending, collected, overdue)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repository.PaymentStats'
      security:
      - BearerAuth: []
      summary: Payment Stats
      tags:
      - Payments
...
  /projects/{project_id}/lots/{lot_id}/contracts/{contract_id}/payments:
    get:
      consumes:
      - application/json
      description: Get payments for a contract
      parameters:
      - description: Contract ID
        in: path
        name: contract_id
        required: true
        type: integer
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: integer
      - description: Lot ID
        in: path
        name: lot_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List Contract Payments
      tags:
      - Payments
  /projects/{project_id}/lots/{lot_id}/contracts/{contract_id}/payments/{payment_id}:
    get:
      consumes:
      - application/json
      description: Get a payment for a specific contract
      parameters:
      - description: Contract ID
        in: path
        name: contract_id
        required: true
        type: integer
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: integer
      - description: Lot ID
        in: path
        name: lot_id
        required: true
        type: integer
      - description: Payment ID
        in: path
        name: payment_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaymentResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get Contract Payment
      tags:
      - Payments
  /projects/{project_id}/lots/{lot_id}/contracts/{contract_id}/payments/{payment_id}/approve:
    post:
      consumes:
      - application/json
      description: Approve a payment for a specific contract
      parameters:
      - description: Contract ID
        in: path
        name: contract_id
        required: true
        type: integer
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: integer
      - description: Lot ID
        in: path
        name: lot_id
        required: true
        type: integer
      - description: Payment ID
        in: path
        name: payment_id
        required: true
        type: integer
      - description: Amount
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.ApprovePaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaymentResponse'
      security:
      - BearerAuth: []
      summary: Approve Contract Payment
      tags:
      - Payments
  /projects/{project_id}/lots/{lot_id}/contracts/{contract_id}/payments/{payment_id}/receipt:
    post:
      consumes:
      - multipart/form-data
      description: Upload receipt for a contract payment
      parameters:
      - description: Contract ID
        in: path
        name: contract_id
        required: true
        type: integer
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: integer
      - description: Lot ID
        in: path
        name: lot_id
        required: true
        type: integer
      - description: Payment ID
        in: path
        name: payment_id
        required: true
        type: integer
      - description: Receipt File
        in: formData
        name: receipt
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Upload Contract Payment Receipt
      tags:
      - Payments
  /projects/{project_id}/lots/{lot_id}/contracts/{contract_id}/payments/{payment_id}/reject:
    post:
      consumes:
      - application/json
      description: Reject a payment for a specific contract
      parameters:
      - description: Contract ID
        in: path
        name: contract_id
        required: true
        type: integer
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: integer
      - description: Lot ID
        in: path
        name: lot_id
        required: true
        type: integer
      - description: Payment ID
        in: path
        name: payment_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaymentResponse'
      security:
      - BearerAuth: []
      summary: Reject Contract Payment
      tags:
      - Payments
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"